requirements.txt
dataclasses_json==0.6.7
datasets==3.6.0
ir_datasets==0.5.10
ir_measures==0.3.7
google-genai==1.24.0
openai==1.91.0
opensearch_py==3.0.0
protobuf==6.31.1
pydantic==2.11.7
python-dotenv==1.1.1
sentence_transformers==4.1.0
setuptools==80.9.0
tiktoken==0.9.0
torch==2.7.1
transformers==4.52.4
